+++ src/components/EventDetails/EventDetails.styles.ts
@@ -23,5 +23,8 @@ export const getStyles = (theme: GrafanaTheme2) => {
         }
       }
     `,
+    description: css`
+      word-break: break-all;
+    `,
   };
 };